virtio-blk driver
As the library, it's pretty messy right now and basically no error
handling and a lot of trust in the caller.

It's kind of working though. It is possible to create a MFS partition,
mount it and copy some files etc. Multithreading should work, but
needs support for indirect descriptors, otherwise the queue gets
full very easily.

virtio_blk.h taken from the Linux sources with types renamed.

Needs a lot of cleanup and restructuring, but it's a start.