VM: Don't put low (4 MB) memory pages into page cache
Observation: Allocating contiguous memory below 16 MB for DMA fails when
using sb16. Although pages are generally allocated from high to low, the
page cache seems to hold a large amount of low memory pages which then
get reused. This is likely because initially modules are loaded < 16 MB
and after relocation, the freed pages end up in the page cache.

Work around this issue by not putting any pages < 4 MB into the page
cache.

This is not a clean fix - it's a hack because:
  1. low memory should probably be explicitly reserved
  2. allocation of low memory does not check the page cache, AFAICS
  3. The 4 MB boundary is hard coded
  4. Only an assumption about why there's no space < 16 MB left
  5. and probably more...

...but it works.